updated un/install script added option to delete only bin or all, remaved keymaps from data.json, added config.json

+++ unistall.sh
@@ -1,12 +1,19 @@
 #! /usr/bin/env bash
 
-# TODO flag to choose deleting data.joson or only bin.
+
+opt1="$1"
 
 user=$(whoami)
 
 dir="/home/$user/openup/"
 
-rm -rf $dir
+if [[ $opt1 == "-a" ]]; then
+        rm -rf $dir
+fi
+if [[ $opt1 == "-h" ]]; then
+        echo -e "\033[0;32mUsage: run {./uninstall.sh} to remove bin. use -a to remove all data\033[0m"
+        exit 0
+fi
 
 sudo rm "/home/$user/.local/bin/openup"
